% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_levels.R
\name{change_levels}
\alias{change_levels}
\title{Change factors by supplying a list of new values}
\usage{
change_levels(x, ...)
}
\arguments{
\item{x}{A factor vector}

\item{...}{Key value pairs for replacing old factor values with new values}
}
\value{
A factor vector
}
\description{
Change factors by supplying a list of new values
}
\examples{
change_levels(as.factor(mtcars$cyl), beast = '8', lame = c('4', '6'))
}
